rename schesmesController.js to schemesController.js
